requirements.txt
django==1.5
redis
django-redis-cache
MySQL-python==1.2.5
python-mimeparse==0.1.4
django-tastypie==0.11.0
django-tastypie-swagger==0.1.2
dj-static==0.0.6
static3==0.5.1

